Add force-port Flag For Use With Mock Mode
For SDK performance testing I want to test each SDK against an instance
of Gocaves running in mock-mode in a Docker container. However
previously using mock-mode would randomly assign ports to each service
meaning I would have to parse the log messages to get the ports before
being able to connect.

Instead I have added a way to force the port that the KV service runs on
so I can use Gocaves for Performance Testing. This means I have had to
add a pointer to an int that permeates down to where the KV port is set,
which isn't the prettiest of ways but the only one I could come up with.
If you can think of a better alternative I am more than happy to change
the implementation.